ID Long -> Int

+++ app/models/Movie.scala
@@ -6,11 +6,11 @@ import services.SparkService
  * Created by yuriy on 2/27/15.
  */
 
-case class Movie(id: Long, title: String, prodDate: String, link: String) extends Serializable
+case class Movie(id: Int, title: String, prodDate: String, link: String) extends Serializable
 
 object Movie {
   def fromString(input: String) = input.split("\\|+").take(4) match {
-    case Array(id, title, date, link) => Movie(id.toLong, title, date, link)
+    case Array(id, title, date, link) => Movie(id.toInt, title, date, link)
   }
 
   def listMovies: List[Movie] = {
